moved error reps into errors mod. removed unused reps
